Moving to the standard Gradle Java project layout
The layout is described [here](https://docs.gradle.org/current/userguide/java_plugin.html#sec:java_project_layout) and makes it so that we need less code in build.gradle because we're using the default locations. Developers familiar with other Gradle (or Maven) projects will be accustomed to seeing them in these locations.